+++ NEWS.md
@@ -6,6 +6,8 @@
 
 **New features**
 
+- Custom tieder `tidy_parameters()` based on `parameters::model_parameters()`
+  (#104)
 - `model_get_coefficients_type()` now returns "prop_hazard" for cloglog-binomial
   models (#106)