+++ NEWS.md
@@ -1,4 +1,4 @@
-# broom.helpers (development version)
+# broom.helpers 1.3.0
 
 **New supported models**